refactor(runtime): flat static platform-handler library, no dynamic import
Replace the 5-step handler resolution waterfall with a transparent
lookup path:

- The platform handler library is now a module-scope
  PLATFORM_HANDLER_LIBRARY map built once from a static
  `import * as` of './handlers', eliminating the runtime's only
  dynamic import (and its documented Jest/Istanbul coverage blind
  spot) plus the category-probing loops that could never match a flat
  `export *` barrel.
- invokeHandler() keeps the exact resolution order and error messages:
  platform.* -> DI platformHandlers then the library; custom -> DI
  customHandlers by full name, then last segment, then a subclass
  method. DI maps stay live (tests register handlers after
  construction), so lookups consult them at call time.
- invokeCustomHandler() drops its `(this as any)[name]` cast for an
  unknown-record narrow.

Handler modules now import Context from '../context' directly instead
of through '../base-mfe', so the static import introduces no cycle.

Refs #253, #254, #262
